feat: Обновить SDK до Remnawave API v2.8.0
Новые эндпоинты:
- GET /tokens/scopes (api_tokens_management.get_scopes)
- GET /users/stream — keyset-пагинация (users.get_users_stream)
- PATCH /hosts/bulk/update (hosts_bulk_actions.update_hosts)
- POST /bandwidth-stats/nodes/users (bandwidthstats.get_stats_nodes_users_usage)

Удалённые эндпоинты (вслед за API):
- POST /hosts/bulk/set-inbound, /hosts/bulk/set-port (заменены update_hosts)
- POST /system/tools/happ/encrypt

Изменения моделей:
- Hosts: tags (вместо tag), mihomoIpVersion, pinnedPeerCertSha256,
  verifyPeerCertByName, finalMask, xhttpExtraParams; fingerprint -> str;
  удалены tag/allowInsecure (с обратной совместимостью через свойства)
- API-токены: name/expiresInDays/scopes, tokens/enabled; новый enum Scope
  (полный каталог scope-ов v2.8.0)
- Nodes: nodeConsumptionMultiplier, note, proxyUrl; тело RestartNodeRequestBodyDto
- trafficLimitBytes int -> float (users, nodes, bulk-actions)
- Infra billing: name, nullable nodeUuid, обязательное nextBillingAt
- HWID: requestIp; webhook-события (user/service) обновлены
- Новый enum MihomoIpVersion

Исправления:
- Bulk-ответы хостов переведены на RootModel (раньше падали с ValueError)
- revoke_user_subscription: исправлен порядок Annotated/Optional (тело терялось)
- Restart*RequestBodyDto: populate_by_name (force_restart игнорировался)

Зависимости:
- Убраны избыточные пины pydantic и pydantic-core (покрыты pydantic[email])
- requires-python <3.14 -> <3.15, добавлен классификатор Python 3.14

Версия пакета 2.7.1 -> 2.8.0.

+++ remnawave/models/users_bulk_actions.py
@@ -38,9 +38,9 @@ class BulkResetTrafficUsersRequestDto(BaseModel):
 class UpdateUserFields(BaseModel):
     """Fields to update for users"""
     status: Optional[UserStatus] = None
-    traffic_limit_bytes: Optional[int] = Field(
-        None, 
-        serialization_alias="trafficLimitBytes", 
+    traffic_limit_bytes: Optional[float] = Field(
+        None,
+        serialization_alias="trafficLimitBytes",
         ge=0,
         description="Traffic limit in bytes. 0 - unlimited"
     )
@@ -98,7 +98,7 @@ class BulkExtendExpirationDateRequestDto(BaseModel):
 class BulkAllUpdateUsersRequestDto(BaseModel):
     """Request to update all users"""
     status: Optional[UserStatus] = Field(default=UserStatus.ACTIVE)
-    traffic_limit_bytes: Optional[int] = Field(
+    traffic_limit_bytes: Optional[float] = Field(
         None,
         serialization_alias="trafficLimitBytes",
         ge=0,
